Moved MapLegend component to legacy components/ folder
Moved images & icons into assets folder
