+++ src/components/MapLegend/IncidentLegendGrid.tsx
