+++ src/components/MapLegend/LegendStyles.tsx
