+++ src/components/MapLegend/assets/construction.png
