+++ src/components/MapLegend/assets/dividers.png
