+++ src/components/MapLegend/assets/footways.png
