+++ src/components/MapLegend/assets/guardrail-legend.png
