+++ src/components/MapLegend/index.ts
@@ -1,7 +1,7 @@
-import ConstructionLegendImage from './construction.png';
-import DividersLegendImage from './dividers.png';
-import FootwaysLegendImage from './footways.png';
-import GuardrailLegendImage from './guardrail-legend.png';
+import ConstructionLegendImage from './assets/construction.png';
+import DividersLegendImage from './assets/dividers.png';
+import FootwaysLegendImage from './assets/footways.png';
+import GuardrailLegendImage from './assets/guardrail-legend.png';
 
 export * from './Legend';
 export * from './IncidentLegendGrid';
